+++ packages/headless/src/controllers/facets/range-facet/numeric-facet/headless-numeric-facet.test.ts
@@ -19,6 +19,7 @@ import {buildMockNumericFacetResponse} from '../../../../test/mock-numeric-facet
 import {buildMockNumericFacetRequest} from '../../../../test/mock-numeric-facet-request';
 import {SearchAppState} from '../../../../state/search-app-state';
 import * as FacetIdDeterminor from '../../_common/facet-id-determinor';
+import {NumericFacetRegistrationOptions} from '../../../../features/facets/range-facets/numeric-facet-set/interfaces/options';
 
 describe('numeric facet', () => {
   const facetId = '1';
@@ -57,7 +58,10 @@ describe('numeric facet', () => {
   });
 
   it('registers a numeric facet with the passed options', () => {
-    const action = registerNumericFacet({facetId, ...options});
+    const action = registerNumericFacet({
+      facetId,
+      ...options,
+    } as NumericFacetRegistrationOptions);
     expect(engine.actions).toContainEqual(action);
   });